requirements.txt
discord.py
emoji
crontab
youtube_dl
PyNaCl